pyproject.toml
[project]
name = "vega-widget"
version = "0.1.0"
description = "Plain vega visualizations as marimo-compatible anywidget"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anywidget[dev]>=0.9.21",
]

[project.optional-dependencies]
ibis = [
    "ibis-framework[duckdb]>=12.0.0",
]
pandas = [
    "pandas>=3.0.1",
]
polars = [
    "polars>=1.39.3",
]

[tool.uv.build-backend]
module-root = "src/py"

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "marimo>=0.21.1",
    "ruff>=0.15.7",
]
